Change student id db type

+++ Sql/create_students.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS students(
-  id     integer      PRIMARY KEY,
+  id     bigserial    PRIMARY KEY,
   name   varchar(500) NOT NULL,
   email  varchar(256) NOT NULL UNIQUE
 );
